+++ src/attestation/reference/snp/mod.rs
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: Apache-2.0
+
+//! SNP launch digest and ID block reference value calculation.
+
+#[cfg(feature = "openssl")]
+pub mod idblock;
+
+pub mod measurement;
+
+pub use measurement::*;