+++ src/lib.rs
@@ -98,23 +98,17 @@ compile_error!(
 /// SEV and SEV-SNP certificates interface.
 pub mod certs;
 
-#[cfg(feature = "snp")]
+#[cfg(any(feature = "sev", feature = "snp"))]
 pub mod attestation;
 
-#[cfg(feature = "snp")]
+#[cfg(any(feature = "sev", feature = "snp"))]
 pub mod snp;
 
 pub mod firmware;
 pub mod launch;
-#[cfg(all(
-    any(feature = "sev", feature = "snp"),
-    any(feature = "openssl", feature = "crypto_nossl")
-))]
-pub mod measurement;
 #[cfg(all(target_os = "linux", feature = "openssl", feature = "sev"))]
 pub mod session;
 mod util;
-pub mod vmsa;
 
 /// Error module.
 pub mod error;