+++ src/snp/mod.rs
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
 
-//! Portable SNP ABI types and helpers.
+//! Portable SNP UAPI types and helpers.
 
 pub mod types;