+++ tests/id-block.rs
@@ -12,12 +12,9 @@ use base64::{engine::general_purpose, Engine as _};
 use hex::{self, FromHex};
 
 use sev::{
-    measurement::{
-        idblock::{load_priv_key, snp_calculate_id},
-        idblock_types::{IdAuth, SevEcdsaPubKey, SevEcdsaSig},
-        snp::SnpLaunchDigest,
-    },
+    attestation::reference::snp::idblock::{load_priv_key, snp_calculate_id},
     parser::{ByteParser, Decoder},
+    snp::types::{IdAuth, SevEcdsaPubKey, SevEcdsaSig, SnpLaunchDigest},
 };
 
 // Testing that the appropriate id-block and key digests are being generated.